[BugFix] Align Contact.zero geom-id dtype with make_data (int64)
_make_data_contact in io.py initializes geom1/geom2/geom as int64, but
Contact.zero() emits int32.  When an env has no contacts (cartpole),
collision() calls Contact.zero() on every step, which overwrites the
int64 init with int32 and triggers a Dynamo recompile on call 2:

  tensor 'contact.geom1' dtype mismatch. expected Long, actual Int

Normalize Contact.zero to int64 for geom fields so the step output's
dtype matches the step input's dtype.

+++ mujoco_torch/_src/types.py
@@ -994,9 +994,9 @@ class Contact(MjTensorClass):
             solreffriction=torch.zeros(shape + (mujoco.mjNREF,), device=device),
             solimp=torch.zeros(shape + (mujoco.mjNIMP,), device=device),
             contact_dim=torch.zeros(shape, dtype=torch.int32, device=device),
-            geom1=torch.zeros(shape, dtype=torch.int32, device=device),
-            geom2=torch.zeros(shape, dtype=torch.int32, device=device),
-            geom=torch.zeros(shape + (2,), dtype=torch.int32, device=device),
+            geom1=torch.zeros(shape, dtype=torch.int64, device=device),
+            geom2=torch.zeros(shape, dtype=torch.int64, device=device),
+            geom=torch.zeros(shape + (2,), dtype=torch.int64, device=device),
             efc_address=torch.zeros(shape, dtype=torch.int64, device=device),
             batch_size=list(shape),
         )
